# perf: ⚡️ remove regex all_white_space