```
regex                   time:   [19.178 ns 19.272 ns 19.423 ns]                   
function                time:   [2.4545 ns 2.4725 ns 2.4976 ns]                      
```